refactor: refactor crawl.py for better maintainability
